Final fixes on the github validation
